Custom mixin - Element width same as Ionic column
This mixin will serve us when we want to add the 'size="12"' or some other value for a custom element (e.g. div). This way, when we add the 'size' attribute on some custom element by implementing this mixin, that element will have same width as the ionic column with the same attribute and value.